backend/requirements.txt
Django==4.2
djangorestframework==3.16.1
django-cors-headers==4.0.0
djangorestframework-simplejwt==5.3.0
gunicorn
setuptools
